melhorando servico user
