+++ app/services/user/IUser.ts
@@ -1,0 +1,7 @@
+export type User = {
+    id: string
+    name: string
+    email: string
+    password: string
+  }
+
